--- a/src/app/modules/angular-slickgrid/models/filterCondition.interface.ts
+++ b/src/app/modules/angular-slickgrid/models/filterCondition.interface.ts
@@ -1,4 +1,5 @@
 import { FilterConditionOption } from './filterConditionOption.interface';
+import { SearchTerm } from './searchTerm.type';
 
 
-export type FilterCondition = (options: FilterConditionOption) => boolean;
+export type FilterCondition = (options: FilterConditionOption, parsedSearchTerms?: SearchTerm | SearchTerm[]) => boolean;