… src/app/modules/angular-slickgrid/filter-conditions/__tests__/booleanFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/collectionSearchFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateEuroFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateEuroShortFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateIsoFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateUsFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateUsShortFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/dateUtcFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/filterConditionProcesses.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/numberFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/objectFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/__tests__/stringFilterCondition.spec.ts
… src/app/modules/angular-slickgrid/filter-conditions/executeMappedCondition.ts
… src/app/modules/angular-slickgrid/services/__tests__/filter.service.spec.ts
… src/app/modules/angular-slickgrid/services/__tests__/graphql.service.spec.ts
… src/app/modules/angular-slickgrid/services/__tests__/grid-odata.service.spec.ts
… src/app/modules/angular-slickgrid/services/__tests__/utilities.spec.ts
… src/app/modules/angular-slickgrid/services/filter.service.ts
… src/app/modules/angular-slickgrid/services/graphql.service.ts
… src/app/modules/angular-slickgrid/services/grid-odata.service.ts
… src/app/modules/angular-slickgrid/services/utilities.ts
… test/cypress/integration/example03.spec.js